Rewrite routes based on resource
Documenting the routes, it's confusing to split
the routes based on POST/GET.
