+++ editor/server.js
@@ -4,7 +4,7 @@ var path = require('path')
 var express = require('express')
 var session = require('express-session')
 var nunjucks = require('nunjucks')
-var routes = require('./app/routes/index.js')
+var routes = require('./app/routes/all.js')
 var documentationRoutes = require('./docs/documentation_routes.js')
 var favicon = require('serve-favicon')
 var app = express()
